Move urlBar and urlBarSuggestions components
Auditors: @bsclifton
